Add DigitalOcean Serverless function for syncing Square products
