+++ .gitignore
@@ -458,3 +458,5 @@ coverage/
 # End of https://www.gitignore.io/api/sonar,sonarqube #
 #######################################################
 .env
+packages/square/product-sync/dist/
+packages/square/product-sync/node_modules/
